# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/callback": {
      "get": {
        "tags": [
          "auth",
          "mvg"
        ],
        "summary": "Callback",
        "operationId": "process_oauth_callback",
        "parameters": [
          {
            "name": "session_state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session State"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth",
          "mvg"
        ],
        "summary": "Refresh Token",
        "operationId": "refresh_access_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/me": {
      "get": {
        "tags": [
          "auth",
          "mvg"
        ],
        "summary": "Read Users Me",
        "operationId": "read_current_user",
        "parameters": [
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/events/upcoming": {
      "get": {
        "tags": [
          "events",
          "mvg"
        ],
        "summary": "Events",
        "operationId": "get_upcoming_events",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  },
                  "title": "Response Get Upcoming Events"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/upcoming_events": {
      "get": {
        "tags": [
          "events",
          "mvg"
        ],
        "summary": "Get Next Event",
        "operationId": "get_next_upcoming_event",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  },
                  "title": "Response Get Next Upcoming Event"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/hash": {
      "post": {
        "tags": [
          "auth",
          "mvg"
        ],
        "summary": "Set Password",
        "operationId": "hash_password",
        "parameters": [
          {
            "name": "pw",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pw"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drinks": {
      "post": {
        "tags": [
          "drinks",
          "mvg"
        ],
        "summary": "Post Drink",
        "operationId": "create_drink",
        "parameters": [
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DrinkCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Drink"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Drink": {
        "properties": {
          "drink": {
            "type": "string",
            "title": "Drink"
          },
          "ml_alc": {
            "type": "integer",
            "title": "Ml Alc"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          },
          "consumer": {
            "$ref": "#/components/schemas/Member"
          },
          "event": {
            "$ref": "#/components/schemas/Event"
          }
        },
        "type": "object",
        "required": [
          "drink",
          "ml_alc",
          "id",
          "time",
          "consumer",
          "event"
        ],
        "title": "Drink"
      },
      "DrinkCreate": {
        "properties": {
          "drink": {
            "type": "string",
            "title": "Drink"
          },
          "ml_alc": {
            "type": "integer",
            "title": "Ml Alc"
          },
          "time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          },
          "consumer_id": {
            "type": "integer",
            "title": "Consumer Id"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          }
        },
        "type": "object",
        "required": [
          "drink",
          "ml_alc",
          "time",
          "consumer_id",
          "event_id"
        ],
        "title": "DrinkCreate"
      },
      "Event": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "duration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration",
            "description": "Duration in Minutes. Null means open end."
          },
          "view_role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Roles"
              }
            ],
            "description": "Minimum role required to view this event",
            "default": "guest"
          },
          "participate_role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Roles"
              }
            ],
            "description": "Minimum role required to participate in this event",
            "default": "mvg-member"
          },
          "edit_role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Roles"
              }
            ],
            "description": "Minimum role required to edit this event",
            "default": "mvg-member"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/Member"
            },
            "type": "array",
            "title": "Participants"
          },
          "author": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Member"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "start_time",
          "location",
          "id",
          "participants",
          "author"
        ],
        "title": "Event"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Member": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "Member"
      },
...
      "Roles": {
        "type": "string",
        "enum": [
          "guest",
          "mkm-member",
          "mvg-member"
        ],
        "title": "Roles"
      },
...
      "TokenRefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "TokenRefreshRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}